momby-admin/backend/python/services/ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-anthropic==0.1.0
openai==1.10.0
anthropic==0.18.1
pinecone-client==3.0.0
chromadb==0.4.22
weaviate-client==4.4.0
sentence-transformers==2.3.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3

